backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0



langchain-core==0.1.45
langchain-groq
langchain-pinecone



httpx==0.27.0
python-dotenv==1.0.1
pytest==8.2.0



youtube-transcript-api
yt-dlp
requests
pinecone
psycopg[binary]
